datacontenedores/python2/requirements.txt
python-socketio[client]
python-socketio[asyncio_client]
fastapi
uvicorn[standard]
celery==5.3.1
Flask==2.0.1
cx_Oracle==8.2.1
schedule==1.1.0
wakeonlan==3.0.0
requests==2.26.0